Added TypeCast support for Float and Boolean.
->Added Support for Boolean and Float in Reader side and Translation Side.

+++ BIR2llvmIR/BIRReader.h
@@ -142,10 +142,10 @@ public:
   ~FloatCpInfo();
 
 private:
-  double value;
+  float value;
 
 public:
-  double getValue() const { return value; }
+  float getValue() const { return value; }
 };
 
 class ByteCpInfo : public ConstantPoolEntry {
